Maintain consistent dtype for pos array

+++ gym_puddle/envs/puddle_env.py
@@ -33,8 +33,10 @@ class PuddleEnv(gym.Env):
         puddles: list[Puddle] | None = None,
         render_mode: str | None = None,
     ):
-        self.start = start if start is not None else np.array([0.2, 0.4])
-        self.goal = goal if goal is not None else np.array([1.0, 1.0])
+        self.start = (
+            start if start is not None else np.array([0.2, 0.4], dtype=np.float32)
+        )
+        self.goal = goal if goal is not None else np.array([1.0, 1.0], dtype=np.float32)
         self.goal_threshold = goal_threshold
         self.noise = noise
         self.thrust = thrust
@@ -48,7 +50,7 @@ class PuddleEnv(gym.Env):
         self.action_space = spaces.Discrete(5)
         self.observation_space = spaces.Box(0.0, 1.0, shape=(2,))
 
-        self.actions = [np.zeros(2) for _ in range(5)]
+        self.actions = [np.zeros(2, dtype=np.float32) for _ in range(5)]
         # [(-0.05, 0), (0.05, 0), (0, -0.05), (0, 0.05), (0, 0)]
         for i in range(4):
             self.actions[i][i // 2] = thrust * (i % 2 * 2 - 1)
